trojan代码架构调整

+++ shell/trojan.sh
@@ -1,128 +1,179 @@
 #!/bin/bash
-
 #curve 不影响已经安装的web服务器,或者想要安装自己的服务器
 #仅占用本域名访问,ip或其他域名访问不影响
+#如果443端口被占用,需要手动修改trojan配置文件端口
 #fonts color
-version=1.1
+version=2.0
 
-yellow(){
+# 脚本格式化输出信息
+function yellow() {
     echo -e "\033[33m\033[01m$1\033[0m"
 }
-green(){
+function green() {
     echo -e "\033[32m\033[01m$1\033[0m"
 }
-red(){
+function red() {
     echo -e "\033[31m\033[01m$1\033[0m"
 }
 
-if [[ -f /etc/redhat-release ]]; then
-    release="centos"
-    systemPackage="yum"
-    systempwd="/usr/lib/systemd/system/"
-elif cat /etc/issue | grep -Eqi "debian"; then
-    release="debian"
-    systemPackage="apt-get"
-    systempwd="/lib/systemd/system/"
-elif cat /etc/issue | grep -Eqi "ubuntu"; then
-    release="ubuntu"
-    systemPackage="apt-get"
-    systempwd="/lib/systemd/system/"
-elif cat /etc/issue | grep -Eqi "centos|red hat|redhat"; then
-    release="centos"
-    systemPackage="yum"
-    systempwd="/usr/lib/systemd/system/"
-elif cat /proc/version | grep -Eqi "debian"; then
-    release="debian"
-    systemPackage="apt-get"
-    systempwd="/lib/systemd/system/"
-elif cat /proc/version | grep -Eqi "ubuntu"; then
-    release="ubuntu"
-    systemPackage="apt-get"
-    systempwd="/lib/systemd/system/"
-elif cat /proc/version | grep -Eqi "centos|red hat|redhat"; then
-    release="centos"
-    systemPackage="yum"
-    systempwd="/usr/lib/systemd/system/"
-fi
+# init 初始化
+function init() {
+    # 验证系统状态是否符合安装条件
+    CHECK=$(grep SELINUX= /etc/selinux/config | grep -v "#")
+    if [ "$CHECK" == "SELINUX=enforcing" ]; then
+        red "======================================================================="
+        red "检测到SELinux为开启状态，为防止申请证书失败，请先重启VPS后，再执行本脚本"
+        red "======================================================================="
+        read -p "是否现在重启 ?请输入 [Y/n] :" yn
+        [ -z "${yn}" ] && yn="y"
+        if [[ $yn == [Yy] ]]; then
+            sed -i 's/SELINUX=enforcing/SELINUX=disabled/g' /etc/selinux/config
+            setenforce 0
+            echo -e "VPS 重启中..."
+            reboot
+        fi
+        exit
+    elif [ "$CHECK" == "SELINUX=permissive" ]; then
+        red "======================================================================="
+        red "检测到SELinux为宽容状态，为防止申请证书失败，请先重启VPS后，再执行本脚本"
+        red "======================================================================="
+        read -p "是否现在重启 ?请输入 [Y/n] :" yn
+        [ -z "${yn}" ] && yn="y"
+        if [[ $yn == [Yy] ]]; then
+            sed -i 's/SELINUX=permissive/SELINUX=disabled/g' /etc/selinux/config
+            setenforce 0
+            echo -e "VPS 重启中..."
+            reboot
+        fi
+        exit
+    fi
 
-function install_trojan(){
-CHECK=$(grep SELINUX= /etc/selinux/config | grep -v "#")
-if [ "$CHECK" == "SELINUX=enforcing" ]; then
-    red "======================================================================="
-    red "检测到SELinux为开启状态，为防止申请证书失败，请先重启VPS后，再执行本脚本"
-    red "======================================================================="
-    read -p "是否现在重启 ?请输入 [Y/n] :" yn
-	[ -z "${yn}" ] && yn="y"
-	if [[ $yn == [Yy] ]]; then
-	    sed -i 's/SELINUX=enforcing/SELINUX=disabled/g' /etc/selinux/config
-            setenforce 0
-	    echo -e "VPS 重启中..."
-	    reboot
-	fi
-    exit
-fi
-if [ "$CHECK" == "SELINUX=permissive" ]; then
-    red "======================================================================="
-    red "检测到SELinux为宽容状态，为防止申请证书失败，请先重启VPS后，再执行本脚本"
-    red "======================================================================="
-    read -p "是否现在重启 ?请输入 [Y/n] :" yn
-	[ -z "${yn}" ] && yn="y"
-	if [[ $yn == [Yy] ]]; then
-	    sed -i 's/SELINUX=permissive/SELINUX=disabled/g' /etc/selinux/config
-            setenforce 0
-	    echo -e "VPS 重启中..."
-	    reboot
-	fi
-    exit
-fi
-if [ "$release" == "centos" ]; then
-    if  [ -n "$(grep ' 6\.' /etc/redhat-release)" ] ;then
-    red "==============="
-    red "当前系统不受支持"
-    red "==============="
-    exit
+    # 初始化包管理
+    if [[ -f /etc/redhat-release ]]; then
+        release="centos"
+        sysPkg="yum"                      # 包管理方式
+        sysPwd="/usr/lib/systemd/system/" # systemctl服务位置
+    elif cat /etc/issue | grep -Eqi "debian"; then
+        release="debian"
+        sysPkg="apt-get"
+        sysPwd="/lib/systemd/system/"
+    elif cat /etc/issue | grep -Eqi "ubuntu"; then
+        release="ubuntu"
+        sysPkg="apt-get"
+        sysPwd="/lib/systemd/system/"
+    elif cat /etc/issue | grep -Eqi "centos|red hat|redhat"; then
+        release="centos"
+        sysPkg="yum"
+        sysPwd="/usr/lib/systemd/system/"
+    elif cat /proc/version | grep -Eqi "debian"; then
+        release="debian"
+        sysPkg="apt-get"
+        sysPwd="/lib/systemd/system/"
+    elif cat /proc/version | grep -Eqi "ubuntu"; then
+        release="ubuntu"
+        sysPkg="apt-get"
+        sysPwd="/lib/systemd/system/"
+    elif cat /proc/version | grep -Eqi "centos|red hat|redhat"; then
+        release="centos"
+        sysPkg="yum"
+        sysPwd="/usr/lib/systemd/system/"
     fi
-    if  [ -n "$(grep ' 5\.' /etc/redhat-release)" ] ;then
-    red "==============="
-    red "当前系统不受支持"
-    red "==============="
-    exit
+
+    # 检测系统最低可容忍版本
+    if [ "$release" == "centos" ]; then
+        if [ -n "$(grep ' 6\.' /etc/redhat-release)" ]; then
+            red "==============="
+            red "当前系统不受支持"
+            red "==============="
+            exit
+        fi
+        if [ -n "$(grep ' 5\.' /etc/redhat-release)" ]; then
+            red "==============="
+            red "当前系统不受支持"
+            red "==============="
+            exit
+        fi
+        systemctl stop firewalld
+        systemctl disable firewalld
+        rpm -Uvh http://nginx.org/packages/centos/7/noarch/RPMS/nginx-release-centos-7-0.el7.ngx.noarch.rpm
+    elif [ "$release" == "ubuntu" ]; then
+        if [ -n "$(grep ' 14\.' /etc/os-release)" ]; then
+            red "==============="
+            red "当前系统不受支持"
+            red "==============="
+            exit
+        fi
+        if [ -n "$(grep ' 12\.' /etc/os-release)" ]; then
+            red "==============="
+            red "当前系统不受支持"
+            red "==============="
+            exit
+        fi
+        systemctl stop ufw
+        systemctl disable ufw
+        apt-get update
     fi
-    systemctl stop firewalld
-    systemctl disable firewalld
-    rpm -Uvh http://nginx.org/packages/centos/7/noarch/RPMS/nginx-release-centos-7-0.el7.ngx.noarch.rpm
-elif [ "$release" == "ubuntu" ]; then
-    if  [ -n "$(grep ' 14\.' /etc/os-release)" ] ;then
-    red "==============="
-    red "当前系统不受支持"
-    red "==============="
-    exit
+}
+
+# install_dependency 安装依赖
+function install_dependency() {
+    green "============================="
+    green "安装依赖项"
+    green "============================="
+    $sysPkg update && $sysPkg upgrade
+    $sysPkg install -y zip tar nginx
+}
+
+# check_domain 验证域名your_name
+function check_domain() {
+    green "============================="
+    green "请输入绑定到本VPS的域名"
+    green "============================="
+    read your_domain
+    real_addr=$(ping ${your_domain} -c 1 | sed '1{s/[^(]*(//;s/).*//;q}')
+    local_addr=$(curl getip.tk)
+    # 如果域名验证失败
+    if [ $real_addr != $local_addr ]; then
+        red "================================"
+        red "域名解析地址与本VPS IP地址不一致"
+        red "本次安装失败，请确保域名解析正常"
+        red "================================"
+        return 1 # 返回码1 域名验证失败
     fi
-    if  [ -n "$(grep ' 12\.' /etc/os-release)" ] ;then
-    red "==============="
-    red "当前系统不受支持"
-    red "==============="
-    exit
+    # 域名验证成功
+    green "=========================================="
+    green "域名解析正常                        "
+    green "=========================================="
+}
+
+# install_cert 安装cert证书
+function install_cert() {
+    #申请https证书
+    mkdir ~/trojan-cert
+    curl https://get.acme.sh | sh -s email=wjuncurve@gmail.com
+    ~/.acme.sh/acme.sh --issue -d $your_domain --webroot /var/www/trojan
+    ~/.acme.sh/acme.sh --installcert -d $your_domain \
+        --key-file ~/trojan-cert/private.key \
+        --fullchain-file ~/trojan-cert/fullchain.cer \
+        --reloadcmd "systemctl force-reload  nginx.service"
+    rm -rf ~/.acme.sh # 删除已无用的acme程序
+
+    if [[ ! -f ~/trojan-cert/fullchain.cer ]]; then
+        red "================================"
+        red "https证书申请失败，本次安装失败"
+        red "================================"
+        rm -rf ~/trojan-cert
+        return 2 #返回码2  cert证书安装失败
     fi
-    systemctl stop ufw
-    systemctl disable ufw
-    apt-get update
-fi
-#$systemPackage -y install  nginx wget unzip zip curl tar >/dev/null 2>&1
-#systemctl enable nginx.service
-apt-get install -y zip tar nginx
-green "======================="
-yellow "请输入绑定到本VPS的域名"
-green "======================="
-read your_domain
-real_addr=`ping ${your_domain} -c 1 | sed '1{s/[^(]*(//;s/).*//;q}'`
-local_addr=`curl getip.tk`
-if [ $real_addr == $local_addr ] ; then
-	green "=========================================="
-	green "       域名解析正常，开始安装trojan"
-	green "=========================================="
-	sleep 1s
-cat > /etc/nginx/sites-enabled/trojan <<-EOF
+    # cert证书安装成功
+    green "=========================================="
+    green "cert证书安装成功                       "
+    green "=========================================="
+}
+
+# install_web 安装web服务
+function install_web() {
+    cat >/etc/nginx/sites-enabled/trojan <<-EOF
 server {
     listen       80;
     server_name  $your_domain;
@@ -137,29 +188,22 @@ server {
     }
 }
 EOF
-#伪站点,位于/var/www/trojan
-mkdir -p /var/www/trojan && cd /var/www/trojan
-wget https://github.com/i-curve/Trojan/raw/master/web.zip && unzip web.zip && rm web.zip
-service nginx restart
-	#申请https证书
-	mkdir ~/trojan-cert && mkdir /etc/trojan
-	curl https://get.acme.sh | sh -s email=wjuncurve@gmail.com
-	~/.acme.sh/acme.sh  --issue  -d $your_domain  --webroot /var/www/trojan
-    	~/.acme.sh/acme.sh  --installcert  -d  $your_domain   \
-        --key-file   ~/trojan-cert/private.key \
-        --fullchain-file ~/trojan-cert/fullchain.cer \
-        --reloadcmd  "systemctl force-reload  nginx.service"
-	if test -s ~/trojan-cert/fullchain.cer; then
-        cd /etc/trojan
-	#wget https://github.com/trojan-gfw/trojan/releases/download/v1.14.0/trojan-1.14.0-linux-amd64.tar.xz
-    wget https://github.com/i-curve/Trojan/raw/master/trojan-1.16.0-linux-amd64.tar.xz
-	tar xf trojan-1.* && rm -f trojan-1.*
-	#下载trojan客户端
-    wget https://github.com/i-curve/Trojan/raw/master/trojan-cli.zip
-	unzip trojan-cli.zip && rm -f trojan-cli.zip
-	cp ~/trojan-cert/fullchain.cer /etc/trojan/trojan-cli/fullchain.cer
-	trojan_passwd=$(cat /dev/urandom | head -1 | md5sum | head -c 8)
-	cat > /etc/trojan/trojan-cli/config.json <<-EOF
+    #伪站点,位于/var/www/trojan
+    mkdir -p /var/www/trojan && cd /var/www/trojan
+    wget https://github.com/i-curve/Trojan/raw/master/web.zip && unzip web.zip && rm web.zip
+    service nginx restart
+}
+
+# install_trojan_client 安装trojan客户端
+function install_trojan_client() {
+    wget https://github.com/i-curve/Trojan/raw/master/trojan-cli.zip # 客户端
+    if [[ ! $? ]]; then                                              # 如果下载失败则直接退出,返回码为3
+        exit 3
+    fi
+    unzip trojan-cli.zip && rm -f trojan-cli.zip
+    cp ~/trojan-cert/fullchain.cer /etc/trojan/trojan-cli/fullchain.cer
+    trojan_passwd=$(cat /dev/urandom | head -1 | md5sum | head -c 8)
+    cat >/etc/trojan/trojan-cli/config.json <<-EOF
 {
     "run_type": "client",
     "local_addr": "127.0.0.1",
@@ -192,8 +236,18 @@ service nginx restart
     }
 }
 EOF
-	rm -rf /etc/trojan/trojan/server.conf
-	cat > /etc/trojan/trojan/server.conf <<-EOF
+}
+
+# install_trojan_server 安装trojan服务端
+function install_trojan_server() {
+    #wget https://github.com/trojan-gfw/trojan/releases/download/v1.14.0/trojan-1.14.0-linux-amd64.tar.xz
+    wget https://github.com/i-curve/Trojan/raw/master/trojan-1.16.0-linux-amd64.tar.xz # 服务端
+    if [[ ! $? ]]; then                                                                # 如果下载失败,返回码为4
+        exit 4
+    fi
+    tar xf trojan-1.* && rm -f trojan-1.*
+    rm -rf /etc/trojan/trojan/server.conf
+    cat >/etc/trojan/trojan/server.conf <<-EOF
 {
     "run_type": "server",
     "local_addr": "0.0.0.0",
@@ -236,14 +290,11 @@ EOF
     }
 }
 EOF
-	cd /etc/trojan/trojan-cli/
-	zip -q -r trojan-cli.zip /etc/trojan/trojan-cli/
-	trojan_path=$(cat /dev/urandom | head -1 | md5sum | head -c 16)
-	mkdir /var/www/html/${trojan_path}
-	mv /etc/trojan/trojan-cli/trojan-cli.zip /var/www/trojan/${trojan_path}/
-	#增加启动脚本
-	
-cat > ${systempwd}trojan.service <<-EOF
+}
+
+# genernate_startup 生成启动脚本
+function genernate_startup() {
+    cat >${sysPwd}trojan.service <<-EOF
 [Unit]  
 Description=trojan  
 After=network.target  
@@ -259,105 +310,101 @@ ExecReload=/bin/kill -HUP \$MAINPID
 [Install]  
 WantedBy=multi-user.target
 EOF
-
-	chmod +x ${systempwd}trojan.service
-	systemctl start trojan.service
-	systemctl enable trojan.service
-	green "======================================================================"
-	green "Trojan已安装完成，请使用以下链接下载trojan客户端，此客户端已配置好所有参数"
-	green "1、复制下面的链接，在浏览器打开，下载客户端"
-	yellow "http://${your_domain}/$trojan_path/trojan-cli.zip"
-	red "请记录下面规则网址"
-	yellow "http://${your_domain}/trojan.txt"
-	green "2、将下载的压缩包解压，打开文件夹，打开start.bat即打开并运行Trojan客户端"
-	green "3、打开stop.bat即关闭Trojan客户端"
-	green "4、Trojan客户端需要搭配浏览器插件使用，例如switchyomega等"
-	green "访问  https://www.v2rayssr.com/trojan-1.html ‎ 下载 浏览器插件 及教程"
-	green "======================================================================"
-	else
-        red "================================"
-	red "https证书没有申请成功，本次安装失败"
-	red "================================"
-	fi
-	
-else
-	red "================================"
-	red "域名解析地址与本VPS IP地址不一致"
-	red "本次安装失败，请确保域名解析正常"
-	red "================================"
-fi
+    chmod +x ${sysPwd}trojan.service
+    systemctl enable trojan.service && systemctl start trojan.service
 }
 
-function remove_trojan(){
+# genernate_download 打包客户端以供下载
+function genernate_download() {
+    cd /etc/trojan/trojan-cli/
+    zip -q -r trojan-cli.zip /etc/trojan/trojan-cli/
+    trojan_path=$(cat /dev/urandom | head -1 | md5sum | head -c 16)
+    mkdir /var/www/html/${trojan_path}
+    mv /etc/trojan/trojan-cli/trojan-cli.zip /var/www/trojan/${trojan_path}/
+}
+
+# install_trojan 安装trojan
+function install_trojan() {
+    install_dependency # 安装依赖项
+    check_domain       # 核对域名
+    install_cert       #申请https证书
+    if [[ "$?" = "2" ]]; then
+        exit 2 # 证书申请失败,则直接退出返回码为: 2
+    fi
+    install_web # 安装web服务
+    #安装trojan
+    mkdir /etc/Trojan && cd /etc/trojan
+    install_trojan_client # 安装trojan客户端
+    install_trojan_server # 安装trojan服务端
+
+    genernate_startup  #增加启动脚本
+    genernate_download # 打包客户端以供下载
+
+    # 安装成功
+    green "======================================================================"
+    green "Trojan已安装完成，请使用以下链接下载trojan客户端，此客户端已配置好所有参数"
+    green "1、复制下面的链接，在浏览器打开，下载客户端"
+    yellow "http://${your_domain}/$trojan_path/trojan-cli.zip"
+    red "请记录下面规则网址"
+    yellow "http://${your_domain}/trojan.txt"
+    green "2、将下载的压缩包解压，打开文件夹，打开start.bat即打开并运行Trojan客户端"
+    green "3、打开stop.bat即关闭Trojan客户端"
+    green "4、Trojan客户端需要搭配浏览器插件使用，例如switchyomega等"
+    green "访问  https://www.v2rayssr.com/trojan-1.html ‎ 下载 浏览器插件 及教程"
+    green "======================================================================"
+}
+
+# remove_trojan 移除trojan
+function remove_trojan() {
     red "================================"
     red "即将卸载trojan"
     red "================================"
-    sleep 1
-    systemctl stop trojan
-    systemctl disable trojan
-    rm -f ${systempwd}trojan.service
-    rm -rf /etc/trojan
-    rm -rf /root/trojan-cert
-    rm -rf /etc/nginx/sites-enabled/trojan
-    rm -rf /var/www/trojan
-    rm -rf ~/.acme.sh/
-    service nginx restart
+
+    systemctl stop trojan && systemctl disable trojan #停止正在运行的trojan服务
+
+    rm -f ${sysPwd}trojan.service          # 删除trojan服务
+    rm -rf /etc/trojan                     # 删除trojan文件
+    rm -rf /root/trojan-cert               # 删除证书
+    rm -rf /etc/nginx/sites-enabled/trojan # 删除nginx中的配置
+    rm -rf /var/www/trojan                 # 删除网站
+    service nginx restart                  # 重启nginx服务
+
     green "=============="
     green "trojan删除完毕"
     green "=============="
 }
-function repair_cert(){
+
+# repair_cert 修复证书过期问题
+function repair_cert() {
     red "================================="
     red "即将修复证书"
     red "================================="
-apt-get install -y zip tar
-green "============================="
-yellow "请输入绑定到本VPS的域名"
-green "============================="
-read your_domain
-real_addr=`ping ${your_domain} -c 1 | sed '1{s/[^(]*(//;s/).*//;q}'`
-local_addr=`curl getip.tk`
-if [ $real_addr == $local_addr ];then
-	green "=========================================="
-	green "       域名解析正常，开始安装trojan"
-	green "=========================================="
-	sleep 1s
+
+    install_dependency # 安装依赖项
+    check_domain       # 核对域名
+
     mv ~/trojan-cert ~/trojan-cert.bake
-	#申请https证书
-	mkdir ~/trojan-cert && mkdir /etc/trojan
-	curl https://get.acme.sh | sh -s email=wjuncurve@gmail.com
-	~/.acme.sh/acme.sh  --issue  -d $your_domain  --webroot /var/www/trojan
-    	~/.acme.sh/acme.sh  --installcert  -d  $your_domain   \
-        --key-file   ~/trojan-cert/private.key \
-        --fullchain-file ~/trojan-cert/fullchain.cer \
-        --reloadcmd  "systemctl force-reload  nginx.service"
-    if test -s ~/trojan-cert/fullchain.cer; then
+    install_cert # 安装https cert证书
+    if [[ "$?" = "2" ]]; then
+        mv ~/trojan-cert.bake ~/trojan-cert
+        exit 2 # 证书安装失败,复原环境并直接退出程序
+    fi
+    rm -rf ~/trojan-cert.bake
+
     cp ~/trojan-cert/fullchain.cer /etc/trojan/trojan-cli/fullchain.cer
     cp ~/trojan-cert/fullchain.cer /var/www/trojan/fullchain.cer
     systemctl reload nginx
-    systemctl stop trojan.service
-    systemctl start trojan.service
-    rm -rf ~/trojan-cert.bake
-    yellow "================================"
-    yellow "安装成功"
-    yellow "可以直接在http://${your_domain}/fullchain.cer 下载cer文件"
+    systemctl restart trojan.service
+
+    green "================================"
+    green "修复成功"
+    green "可以直接在http://${your_domain}/fullchain.cer 下载cer文件"
     red "================================"
-    else
-    red "================================"
-	red "https证书没有申请成功，本次安装失败"
-	red "================================" 
-    rm -rf ~/trojan-cert 
-    mv ~/trojan-cert.bake ~/trojan-cert
-    fi
-else
-	red "================================"
-	red "域名解析地址与本VPS IP地址不一致"
-	red "本次安装失败，请确保域名解析正常"
-	red "================================"
-fi
 }
-start_menu(){
-    clear
+
+# start_menu 脚本入口
+function start_menu() {
+    init && clear # 执行脚本初始化和清屏
     green " ===================================="
     green ' Author:curve'
     green " Trojan 一键安装自动脚本      "
@@ -374,26 +421,26 @@ start_menu(){
     yellow " 0. 退出脚本"
     red " ===================================="
     echo
+
     read -p "请输入数字:" num
     case "$num" in
     1)
-    install_trojan
-    ;;
+        install_trojan
+        ;;
     2)
-    remove_trojan
-    ;;
+        remove_trojan
+        ;;
     3)
-    repair_cert
-    ;;
+        repair_cert
+        ;;
     0)
-    exit 1
-    ;;
+        exit 0
+        ;;
     *)
-    clear
-    red "请输入正确数字"
-    sleep 1s
-    start_menu
-    ;;
+        red "请输入正确数字"
+        sleep 2s
+        start_menu
+        ;;
     esac
 }
